requirements.txt
Flask==3.1.0
llama_index==0.12.11
python-dotenv==1.0.1
Requests==2.32.3
stripe==11.4.1
Werkzeug==3.1.3
